+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.??.?? (????-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed soft transitions with array parameters `base/b-router`
+
 ## v3.30.1 (2022-10-25)
 
 #### :bug: Bug Fix